+++ sportorg/gui/dialogs/teamwork_properties.py
@@ -46,7 +46,6 @@ from sportorg import settings
 from sportorg.gui.global_access import GlobalAccess
 from sportorg.gui.utils.custom_controls import AdvSpinBox
 from sportorg.language import translate
-from sportorg.models.memory import race
 from sportorg.modules.teamwork.crypto import generate_teamwork_key
 from sportorg.modules.teamwork.teamwork import Teamwork
 
@@ -92,6 +91,7 @@ class TeamworkPropertiesDialog(QDialog):
         self.teamwork_groupbox_layout = QFormLayout()
         self.teamwork_item_client = QRadioButton(translate("Client"))
         self.teamwork_item_server = QRadioButton(translate("Server"))
+        self.teamwork_item_autorun = QCheckBox(translate("Run at startup"))
         self.teamwork_groupbox_layout.addRow(self.teamwork_item_client)
         self.teamwork_groupbox_layout.addRow(self.teamwork_item_server)
         self.teamwork_groupbox.setLayout(self.teamwork_groupbox_layout)
@@ -103,6 +103,7 @@ class TeamworkPropertiesDialog(QDialog):
             QLabel(translate("Encryption key")), self.teamwork_key_editor_container
         )
         self.teamwork_layout.addRow(self.teamwork_groupbox)
+        self.teamwork_layout.addRow(self.teamwork_item_autorun)
         self.teamwork_tab.setLayout(self.teamwork_layout)
 
         self.tab_widget.addTab(self.teamwork_tab, translate("Client/Server"))
@@ -264,40 +265,45 @@ class TeamworkPropertiesDialog(QDialog):
         self.refresh_connected_clients()
 
     def set_values_from_model(self):
-        obj = race()
-
-        teamwork_host = obj.get_setting("teamwork_host", "localhost")
-        teamwork_port = obj.get_setting("teamwork_port", 50010)
-        teamwork_type_connection = obj.get_setting("teamwork_type_connection", "client")
-        teamwork_encryption_enabled = obj.get_setting(
-            "teamwork_encryption_enabled", False
+        teamwork_host = str(settings.SETTINGS.teamwork_host or "localhost")
+        try:
+            teamwork_port = int(settings.SETTINGS.teamwork_port)
+        except (TypeError, ValueError):
+            teamwork_port = 50010
+        teamwork_type_connection = str(
+            settings.SETTINGS.teamwork_type_connection or "client"
         )
+        teamwork_encryption_enabled = bool(
+            settings.SETTINGS.teamwork_encryption_enabled
+        )
+        teamwork_autorun = bool(settings.SETTINGS.teamwork_autorun)
 
         self.teamwork_item_host.setText(teamwork_host)
         self.teamwork_item_port.setValue(teamwork_port)
         self.teamwork_item_encryption_enabled.setChecked(teamwork_encryption_enabled)
+        self.teamwork_item_autorun.setChecked(teamwork_autorun)
         self.teamwork_item_encryption_key.setText(
             str(settings.SETTINGS.teamwork_encryption_key or "")
         )
-        if teamwork_type_connection == "client":
-            self.teamwork_item_client.setChecked(True)
-        elif teamwork_type_connection == "server":
+        if teamwork_type_connection == "server":
             self.teamwork_item_server.setChecked(True)
+        else:
+            self.teamwork_item_client.setChecked(True)
 
     def apply_changes_impl(self):
-        obj = race()
-
         teamwork_host = self.teamwork_item_host.text()
         teamwork_port = self.teamwork_item_port.value()
         teamwork_encryption_enabled = self.teamwork_item_encryption_enabled.isChecked()
+        teamwork_autorun = self.teamwork_item_autorun.isChecked()
         teamwork_type_connection = "client"
         if self.teamwork_item_server.isChecked():
             teamwork_type_connection = "server"
 
-        obj.set_setting("teamwork_host", teamwork_host)
-        obj.set_setting("teamwork_port", teamwork_port)
-        obj.set_setting("teamwork_type_connection", teamwork_type_connection)
-        obj.set_setting("teamwork_encryption_enabled", teamwork_encryption_enabled)
+        settings.SETTINGS.teamwork_host = teamwork_host
+        settings.SETTINGS.teamwork_port = teamwork_port
+        settings.SETTINGS.teamwork_type_connection = teamwork_type_connection
+        settings.SETTINGS.teamwork_encryption_enabled = teamwork_encryption_enabled
+        settings.SETTINGS.teamwork_autorun = teamwork_autorun
 
         settings.SETTINGS.teamwork_encryption_key = (
             self.teamwork_item_encryption_key.text()